Added MAC address parsing
Added an API function call to parse MAC address in the form
xx-xx-xx-xx-xx-xx, and to convert IP and MAC address to strings for
output. Also, the component parameters were modified to use MAC strings
rather than a comma delimited list, and the initialization function was
updated to parse the MAC address rather than force load the array data
directly.