Started work incorporating W5500
Started to incorporate the W5500 interface in to the driver.  Fixed
issue during read with fast clock rates that was causing ChipRead() to
read data and return before the data was sent.

W5500 low-level interface was integrated, and socket open and server
mode seem to function. Read/Write is currently broken for the W5500
implementation.